refactor: Alterar atributo da classe Quantidade de BigDecimal para int
No banco de dados este atributo na tabela é declarado como INT,
Então para facilitar e eu não ter que ficar alterando os atributos do banco
mudei o atributo de BigDecimal, que seria equivalente a DECIMAL,
para int que é o tipo correto ao que está no meu banco de dados.

+++ src/main/java/model/venda/ItemVenda.java
@@ -12,14 +12,14 @@ public class ItemVenda {
     public ItemVenda(Produto produto, Quantidade quantidade) {
         this.produto = produto;
         this.quantidade = quantidade;
-        this.total = new Total(produto, this.quantidade.getQuantidade());
+        this.total = new Total(produto, BigDecimal.valueOf(this.quantidade.getQuantidade()));
     }
 
     int getProdutoId() {
         return this.produto.getProdutoId();
     }
 
-    BigDecimal getQuantidade() {
+    int getQuantidade() {
         return this.quantidade.getQuantidade();
     }
